requirements.txt
anthropic>=0.25.0
web3>=7.0.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
